backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0
tensorflow==2.15.0
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
slowapi==0.1.9
python-json-logger==2.0.7
